better use of localonly test

+++ test/commands/shane/github/githubSourceInstall.test.ts
@@ -12,30 +12,33 @@ const username = 'mshanemc';
 
 describe('shane:github:src:install', () => {
 
-  before(async () => {
-    await exec(`rm -rf ${testProjectName}`);
-    await exec(`sfdx force:project:create -n ${testProjectName}`);
-    await testutils.orgCreate(testProjectName);
-  });
+  if (!process.env.LOCALONLY) {
+    before(async () => {
+      await exec(`rm -rf ${testProjectName}`);
+      await exec(`sfdx force:project:create -n ${testProjectName}`);
+      await testutils.orgCreate(testProjectName);
+    });
 
-  it('works with specified src folder', async () => {
+    it('works with specified src folder', async () => {
 
-    const repo = 'lightningErrorHandler';
+      const repo = 'lightningErrorHandler';
 
-    const results = await exec(`sfdx shane:github:src:install -g ${username} -r ${repo} -p src --json`, { cwd: testProjectName });
+      const results = await exec(`sfdx shane:github:src:install -g ${username} -r ${repo} -p src --json`, { cwd: testProjectName });
 
-    // console.log(results);
-    expect(results).to.be.an('object');
-    expect(results.stdout).to.be.a('string');
-    const stdout = JSON.parse(results.stdout);
-    // console.log(stdout);
-    // console.log(stdout.status);
-    expect(stdout.status).to.equal(0);
+      // console.log(results);
+      expect(results).to.be.an('object');
+      expect(results.stdout).to.be.a('string');
+      const stdout = JSON.parse(results.stdout);
+      // console.log(stdout);
+      // console.log(stdout.status);
+      expect(stdout.status).to.equal(0);
 
-  });
+    });
 
-  after(async () => {
-    await testutils.orgDelete(testProjectName);
-    await exec(`rm -rf ${testProjectName}`);
-  });
+    after(async () => {
+      await testutils.orgDelete(testProjectName);
+      await exec(`rm -rf ${testProjectName}`);
+    });
+  }
+
 });
